Version0.9.2 TiS Skins and Fixes
[h3]Together In Spire[/h3]
- Shadow, Tails, Knuckles skins and respective colored cardbacks
- Fix Potions

[h3]Fixes[/h3]
Chao Garden Event - The vanilla cast and the Sonic cast should be available as drink buddies again.
Rouge Event - You can talk to Omochao to remember what Assists and DO NOT THROW mean.

[h3]Balance[/h3]
Loop De Loop - Exhaustive 0 -> Exhaustive 4. I don't think achieving an infinite is that easy, but I'm just one guy.

[h3]QoL[/h3]
Enable Cross Mod Integration - This is a new setting and it's very untested. If your game is buggy and crashing often, this is a last ditch setting to try to fix things.